backend/requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
alabaster==1.0.0
attrs==24.2.0
babel==2.16.0
bitsandbytes==0.44.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
datasets==3.1.0
dill==0.3.8
docopt==0.6.2
docutils==0.21.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.9.0
gTTS==2.5.4
huggingface-hub==0.26.2
idna==3.10
imagesize==1.4.1
# indic_nlp_library_IT2 @ git+https://github.com/VarunGumma/indic_nlp_library@601521e05ed0ed8f2165ac317a47d186e25b6f0d
# Editable install with no version control (IndicTransToolkit==1.0.2)
# -e /home/nagajas/Desktop/MINIPROJ02/DS510-MiniProject2/IndicTrans2/huggingface_interface/IndicTransToolkit
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
mock==5.1.0
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
openfile==0.0.7
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==3.0.0
propcache==0.2.0
psutil==6.1.0
pyarrow==18.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacrebleu==2.4.3
sacremoses==0.1.1
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.5.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.20.3
toolwrapper==2.1.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
uctools==1.3.0
urllib3==2.2.3
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.17.2
